backend/requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==2.5.1
google-auth==2.21.0
google-auth-oauthlib==1.0.0
google-api-python-client==2.94.0
openai==0.27.8
python-dotenv==1.0.0
